+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `useProductDispatch` hook.
+
+### Fixed
+- TypeScript types not being picked-up and exported correctly by React builder.
 
 ## [0.8.1] - 2020-07-22
 ### Fixed